Moves the option processing out of the command
The logic for option processing was cluterring the command file.
While it was clear that that code should be moved elsewhere, the
question was where it should be put.

The proposed solution is to change the folder structure to
accomodate command related files. Ideally the command itself should
be moved there but that change will break compatibility.

For now the Options class was created to handle all option specific
processing and the command will remain untouched.