Adds test utility files
Currently there's low visibility regarding test coverage and we don't
have an easy way to build and test locally. To solve this problem,
it's safe to assume that docker and docker-compose are fairly common
tools so this solution makes use of them.

I added files for build and test and a Makefile to make things easier
and quicker. I also updated phpunit.xml to adjust for coverage (it
was considering the tests folder and not the src folder), and added
instruction to the readme on how to use the new utilities.