Fixes an error that would cause the command fail
The PhpKafkaConsumerCommand::handle() method validates and passes along
options given by using options as a propety and not as a method. There
is no options property. The correct way to get all option given to the
command is to call the method options().

This issue has been fixes simply by adding the missing parenthesis.