Adds integration tests and dev conveniences
It is arguably easier to contribute to a project and to decide to use a
library or not based on its tests: are they well written? Up to date? Is
there proper coverage? Testing is an essential part of assunring the
quality of the project.

In this sense, to help improve the quality of the project integration
tests were added as well as utilities for building and testing locally
with Docker, Docker Compose, and Make.
These additions will not be exported in the releases as specified in the
composer.json file in the section "archive". Finally, the readme.md file
has been updated to include build and test instrctions.