Version test and related fixes (#16)
* Adds github actions file

* Adds files for version tests

As mentioned in #11, we need an easy way to state our coverage regarding
the main dependencies: PHP, librdkafka, ext-rdkafka, and Laravel.

This commit adds Docker files for building containers to run this
dependency test and give us result directly from Github Actions.

* Removes unecessary method call in test setup

The call to Dotenv::create() in the setUp() method for the integration
tests was breaking in new version of Laravel and it was unecessary.

This call was removed and we should see improvement in the version
tests.

* Removes call to Conf::setDefaultTopicConf()

This call has been deprecated and the documentation for ext-rdkafka
states: "This function has been DEPRECATED. Relying on this feature
is highly discouraged. Set default topic settings normally like
global configuration settings."

As recommended, the default config has been moved to the global
configuration settings.