Removes unecessary files
The docker-compose-test.yaml and its associated commands on Makefile
were no longer needed. They should have been removed before the merge of
\#16 but they weren't.

They are been removed now. In order to perform the version tests, we can
use the numbered tests in Makefile with the associated files in the
build/docker-compose-files directory.