Extends tests coverage
Before the next release, we should make sure that the changes added in
the last commits didn't break anything. The most used metric for that is
test coverage.

This commit extends test coverage. While we haven't reached one hundred
percent, the main point are well covered and all the other missing
points can be covered in a later release.