Makes tests wait for Kafka container
The depends_on directive is not enough to ensure that the version and
integration tests will run only after the container running Kafka are
ready to handle requests, which causes random errors in our pipeline due
to topics not being available.

The solution is to add a wait-for-it script shamelessly copied from
https://github.com/vishnubob/wait-for-it. The use of that script in
start.sh should be enough to ensure our tests run in a consistent
environment all the time.