Create event dispatcher
Implement the concept of events and create a event dispatcher to
dispatch events.

Create the MessageHandled event and dispatch it when a  message is handled
successfully. Also, enable the subscriber to stop message consumption
from kafka through this event. This allows for the users to
implement some sort of graceful shutdown.