Fix internal retry rdkafka error.
We made an adjustment to disregard the error RD_KAFKA_RESP_ERR__TIMED_OUT which is internal to the lib of rdkafka.
This error happens because we don't have a message to be consumed and the return is null. Only that the error is registered in rdkafka, which is not necessarily an error that we must consider.