Improve kafka config handling
Right now is not possible to customize consumer options, making it
impossible for more advanced use cases that need a fine-grained control
of the consumer options.

Add possibility to use auto commit instead of the default manual commit.

Add possibility for the user to pass custom kafka options to the
consumer.