Fixes error caused by undefined property
The command options class had a reference to the property config which
belongs to the command and was not passed to the options when we
separated the opptions into their own class.

In specific situations where the Laravel command would be invoked
without a group id option, this would cauise an error.

We removed the reference to the undefined property and made the command
check what group id value it should pass to the options constructor.

Fixes #44